Changed type of ports to uint16_t. Added interpolation of min/max ports.

+++ src/server.h
@@ -4,7 +4,7 @@
 // by default, serve on all interfaces
 #define SERVER_ADDR 0x0000
 
-bool readPorts(char** argv, int* ports);
+bool readPorts(char** argv, uint16_t* ports);
 void serve(uint16_t engPort, uint16_t maoPort, uint16_t gerPort);
 
 #endif
